requirements.txt
Flask
Pillow
rq
redis